Add drop-all-tables command
